задание 7-2
